Building from Scratch with Docker
A developer is working on a project that utilizes Docker for containerization.
They have been making updates to the project and need to rebuild the containers
from scratch to ensure all changes are reflected. They discover a new flag,
"--build-from-scratch", that allows them to do this. They update the code to
include this flag and successfully rebuild their containers from scratch.